+++ schema/0004_page_title.up.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "hit" ADD COLUMN "title" String;
+ALTER TABLE "event" ADD COLUMN "title" String;